conf: move MakeUrl to package conf
Avoid cyclic imports
