+++ conf/util_test.go
@@ -6,14 +6,17 @@
 // modification, are permitted under the terms of the BSD License. See
 // LICENSE file in the root of the Project.
 
-package web
+package conf
 
-import (
-	"os"
-	"testing"
-)
+import "testing"
 
-func TestMain(m *testing.M) {
-	os.Chdir("..")
-	os.Exit(m.Run())
+func TestMakeUrl(t *testing.T) {
+	u := MakeUrl("/foo/%d", 200)
+	if u != "http://localhost:8080/foo/200" {
+		t.Error("Wrong url")
+	}
+	u = MakeUrl("/foo/%s", "some string")
+	if u != "http://localhost:8080/foo/some+string" {
+		t.Error("Wrong url")
+	}
 }
